Fix crash bugs: remove dead WorkManager Hilt binding, fix marketCap type, add Room migration fallback
- Remove provideWorkManager from AppModule: it was dead code providing WorkManager
  as a Hilt singleton, posing an initialization-order crash risk if triggered before
  HiltWorkerFactory injection completes
- Change CoinMarketDto.marketCap from Long? to Double?: CoinGecko API returns market
  cap as a floating-point JSON number; Gson fails to deserialize it into Long at runtime
  causing a JsonSyntaxException crash when loading coin data
- Update all Coin mapping calls from 0L to 0.0 for marketCap default values
- Update Coin.marketCap domain model field from Long to Double to match
- Add fallbackToDestructiveMigration() to Room database builder to prevent
  IllegalStateException on schema mismatch during fresh installs
- Suppress deprecated window.statusBarColor warning for API 35 compatibility

+++ app/src/main/java/com/cryptotracker/data/remote/dto/CoinMarketDto.kt
@@ -8,7 +8,7 @@ data class CoinMarketDto(
     val name: String,
     val image: String,
     @SerializedName("current_price") val currentPrice: Double?,
-    @SerializedName("market_cap") val marketCap: Long?,
+    @SerializedName("market_cap") val marketCap: Double?,
     @SerializedName("market_cap_rank") val marketCapRank: Int?,
     @SerializedName("total_volume") val totalVolume: Double?,
     @SerializedName("high_24h") val high24h: Double?,
